Split off test of LookupNames
